Recompile all test programs

+++ test/dostest/fpipe.c
@@ -5,7 +5,7 @@
 int main()
 {
 	long ret;
-	int hdl[2];
+	short hdl[2];
 
 	ret = Fpipe(hdl);
 	printf("\r\nret = %ld hdl1 = %d hdl2 = %d\n", ret, hdl[0], hdl[1]);
